Moved a bunch of components into @backstage/catalog-react
