+++ plugins/catalog-react/src/components/IconLink/IconLink.tsx
