+++ plugins/catalog-react/src/components/LinksGridList/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/** @public */
-export type {
-  Breakpoint,
-  ColumnBreakpoints,
-} from '@backstage/plugin-catalog-react';
+export { LinksGridList } from './LinksGridList';
