+++ plugins/catalog-react/src/components/index.ts
@@ -28,3 +28,5 @@ export * from './InspectEntityDialog';
 export * from './UnregisterEntityDialog';
 export * from './UserListPicker';
 export * from './EntityProcessingStatusPicker';
+export * from './AboutField';
+export * from './LinksGridList';
