--- plugins/catalog/src/components/AboutCard/AboutContent.tsx
@@ -1,55 +1,0 @@
-/*
- * Copyright 2020 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { Entity } from '@backstage/catalog-model';
-import { Grid } from '@material-ui/core';
-import React from 'react';
-import {
-  DescriptionAboutField,
-  OwnerAboutField,
-  DomainAboutField,
-  SystemAboutField,
-  ParentComponentAboutField,
-  TypeAboutField,
-  LifecycleAboutField,
-  TagsAboutField,
-  LocationTargetsAboutField,
-} from './fields';
-
-/**
- * Props for {@link AboutContent}.
- *
- * @public
- */
-export interface AboutContentProps {
-  entity: Entity;
-}
-
-/** @public */
-export function AboutContent(props: AboutContentProps) {
-  return (
-    <Grid container>
-      <DescriptionAboutField />
-      <OwnerAboutField />
-      <DomainAboutField />
-      <SystemAboutField />
-      <ParentComponentAboutField />
-      <TypeAboutField />
-      <LifecycleAboutField />
-      <TagsAboutField />
-      <LocationTargetsAboutField />
-    </Grid>
-  );
-}
