+++ plugins/catalog/src/components/AboutCard/index.ts
@@ -16,10 +16,3 @@
 
 export { AboutCard } from './AboutCard';
 export type { AboutCardProps } from './AboutCard';
-export { AboutContent } from './AboutContent';
-export type { AboutContentProps } from './AboutContent';
-export { AboutField } from './AboutField';
-export type { AboutFieldProps } from './AboutField';
-
-export * from './buttons';
-export * from './fields';
